+++ package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "cors": "^2.8.5",
-    "next": "12.1.6",
+    "next": "^12.1.7-canary.45",
     "react": "18.1.0",
     "react-dom": "18.1.0"
   },
@@ -20,7 +20,7 @@
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.5",
     "eslint": "8.16.0",
-    "eslint-config-next": "12.1.6",
+    "eslint-config-next": "^12.1.7-canary.45",
     "typescript": "4.7.2"
   }
 }
